♻️ Move the ProtectedRoute component in a separate file
The App component is a little bit too bloated with all those router related code.
